Changing templates to .ejs and streamline
I've changed the template planguage from jade to .ejs (which is plain HTML + JS and so less esoteric and easier for most people to work with).

I've improved layout and copy of pages at the same time.

I've removed auth provider information for services like Instagram and Linked In, and limited it to just sign-on services that are most commonly used (Facebook, Google, Twitter and GitHub).

I've removed the API examples as I have other plans for that.

I've also removed some of the hard coded site values (e.g. the site name, the email address used for automated emails) put them in a configuration file.

This is an initial makeover to suit the repo for future work I have in mind.